Fixed some warning messages like "type convert".Add version for the libpcom library.Redirectory the output folder for QtCreator building.
git-svn-id: svn://192.168.2.188/qtcms@1060 894f9759-fa2a-45cf-9159-0a04a166245e

+++ src/libpcom/libpcom.cpp
@@ -63,7 +63,7 @@ char * pcomGUID2String(const GUID &guid)
 	static char sRet[64] = {0};
     QString sGuid;
     sGuid.sprintf("%08X-%04X-%04X-%02X%02X-%02X%02X%02X%02X%02X%02X",
-                  guid.Data1,
+                  (unsigned int)guid.Data1,
                   guid.Data2,
                   guid.Data3,
                   *((unsigned char *)guid.Data4),
